pyproject.toml
[tool.poetry]
name = "post_service"
version = "0.1.0"
description = ""
authors = ["Denysenko Danylo"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2022.7.1"
python-slugify = "^8.0.0"
Pillow = "^9.4.0"
argon2-cffi = "^21.3.0"
whitenoise = "^6.3.0"
redis = "^4.4.2"
hiredis = "^2.2.1"
Django = "^4.1.6"
django-environ = "^0.9.0"
django-model-utils = "^4.3.1"
django-allauth = "^0.52.0"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"
django-redis = "^5.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.25.1"
Werkzeug = {extras = ["watchdog"], version = "^2.2.2"}
ipdb = "^0.13.11"
mypy = "^0.991"
django-stubs = "^1.14.0"
pytest = "^7.2.1"
pytest-sugar = "^0.9.6"
djangorestframework-stubs = "^1.8.0"
Sphinx = "^6.1.3"
sphinx-autobuild = "^2021.3.14"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"
coverage = "^7.1.0"
black = "^23.1.0"
pylint-django = "^2.5.3"
pre-commit = "^3.0.3"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
django-coverage-plugin = "^3.0.0"
pytest-django = "^4.5.2"
gunicorn = "^20.1.0"
django-anymail = "^9.0"
psycopg2-binary = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
manage = 'manage:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
